requirements.txt
elasticsearch
psycopg2
fastparquet
kafka-python